Move 'Find Contacts' upgrader steps from AdminUI to SearchUI
