+++ ext/civicrm_search_ui/info.xml
@@ -39,4 +39,5 @@
     <mixin>setting-php@1.0.0</mixin>
     <mixin>smarty-v2@1.0.1</mixin>
   </mixins>
+  <upgrader>CRM_CivicrmSearchUi_Upgrader</upgrader>
 </extension>
